Enable multiple objects
Put in the structure for multiple shapes. They will
allocate directly into the openGL buffers for vertices,
indices, and colours but keep pointers to their specific
data. This means that they can be transformed
independently of eachother but keeps the memory nice and
contiguous.

Created a helper method for creating pyramids, will create
methods for different shapes moving forward.