Being implementation of loading .gltf files
Implemented a very hardcoded version of loading .gltf
files. Currently you manually set the size of the vertex and
index buffers and give it a file and it will load the mesh
and render it. This is incredibly fragile and needs to be
made much more robust. Once that has been done we can look
at loading and mapping the textures as well.

To achieve this we needed to include a JSON library for parsing
the .gltf format and a base64 library for decoding the .gltf
buffers.

Forgive the terrible mess, its really late. I'll clean it up
next time.