Implement texture loading and mapping for models
Implemented loading a model's texture from the gltf file
and mapping the texture UVs for that model. The texture
data was encoded as a base64 string inside of the gltf file
so we needed to decode the string and load it with stbi image.
Note that texture PNG that blockbench actually saves to will
NOT work as the texture file(I think some cropping/packing
happens before it is exported.